# src/main/java/com/josue/kodeur/xtremanalyse/api/api-docs.json
{
  "openapi": "3.0.1" ,
  "info": {
    "title": "OpenAPI definition, Xtrem-Analyse" ,
    "version": "v0"
  } ,
  "servers": [
    {
      "url": "http://localhost:9000" ,
      "description": "Generated server url"
    }
  ] ,
  "paths": {
...
    "/api/v1/personne-impliquee/update/{typePersonne}/{professionId}/{immatriculation}/{accidentId}": {
      "put": {
        "tags": [
          "personne-impliquee-controller"
        ] ,
        "operationId": "update_3" ,
        "parameters": [
          {
            "name": "typePersonne" ,
            "in": "path" ,
            "required": true ,
            "schema": {
              "type": "string"
            }
          } ,
          {
            "name": "professionId" ,
            "in": "path" ,
            "required": true ,
            "schema": {
              "type": "integer" ,
              "format": "int64"
            }
          } ,
          {
            "name": "immatriculation" ,
            "in": "path" ,
            "required": true ,
            "schema": {
              "type": "string"
            }
          } ,
          {
            "name": "accidentId" ,
            "in": "path" ,
            "required": true ,
            "schema": {
              "type": "integer" ,
              "format": "int64"
            }
          }
        ] ,
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonneImpliquee"
              }
            }
          } ,
          "required": true
        } ,
        "responses": {
          "400": {
            "description": "Bad Request" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "401": {
            "description": "Unauthorized" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "200": {
            "description": "OK" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PersonneImpliquee"
                }
              }
            }
          }
        }
      }
    } ,
    "/api/v1/assurance/update": {
      "put": {
        "tags": [
          "assurance-controller"
        ] ,
        "operationId": "update_4" ,
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Assurance"
              }
            }
          } ,
          "required": true
        } ,
        "responses": {
          "400": {
            "description": "Bad Request" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "401": {
            "description": "Unauthorized" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "200": {
            "description": "OK" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Assurance"
                }
              }
            }
          }
        }
      }
    } ,
...
    "/api/v1/personne-impliquee/add/{typePersonne}/{professionId}/{immatriculation}/{accidentId}": {
      "post": {
        "tags": [
          "personne-impliquee-controller"
        ] ,
        "operationId": "add_4" ,
        "parameters": [
          {
            "name": "typePersonne" ,
            "in": "path" ,
            "required": true ,
            "schema": {
              "type": "string"
            }
          } ,
          {
            "name": "professionId" ,
            "in": "path" ,
            "required": true ,
            "schema": {
              "type": "integer" ,
              "format": "int64"
            }
          } ,
          {
            "name": "immatriculation" ,
            "in": "path" ,
            "required": true ,
            "schema": {
              "type": "string"
            }
          } ,
          {
            "name": "accidentId" ,
            "in": "path" ,
            "required": true ,
            "schema": {
              "type": "integer" ,
              "format": "int64"
            }
          }
        ] ,
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonneImpliquee"
              }
            }
          } ,
          "required": true
        } ,
        "responses": {
          "400": {
            "description": "Bad Request" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "401": {
            "description": "Unauthorized" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "200": {
            "description": "OK" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PersonneImpliquee"
                }
              }
            }
          }
        }
      }
    } ,
    "/api/v1/assurance/add": {
      "post": {
        "tags": [
          "assurance-controller"
        ] ,
        "operationId": "add_5" ,
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Assurance"
              }
            }
          } ,
          "required": true
        } ,
        "responses": {
          "400": {
            "description": "Bad Request" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "401": {
            "description": "Unauthorized" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "200": {
            "description": "OK" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Assurance"
                }
              }
            }
          }
        }
      }
    } ,
...
    "/api/v1/personne-impliquee/personne-dans-vehicule/{id}/{id2}": {
      "get": {
        "tags": [
          "personne-impliquee-controller"
        ] ,
        "operationId": "personneImpliqueeListByAccident" ,
        "parameters": [
          {
            "name": "id" ,
            "in": "path" ,
            "required": true ,
            "schema": {
              "type": "string"
            }
          } ,
          {
            "name": "id2" ,
            "in": "path" ,
            "required": true ,
            "schema": {
              "type": "integer" ,
              "format": "int64"
            }
          }
        ] ,
        "responses": {
          "400": {
            "description": "Bad Request" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "401": {
            "description": "Unauthorized" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "200": {
            "description": "OK" ,
            "content": {
              "*/*": {
                "schema": {
                  "type": "array" ,
                  "items": {
                    "$ref": "#/components/schemas/PersonneImpliquee"
                  }
                }
              }
            }
          }
        }
      }
    } ,
    "/api/v1/personne-impliquee/personne-dans-accident/{id}": {
      "get": {
        "tags": [
          "personne-impliquee-controller"
        ] ,
        "operationId": "personneImpliqueeList" ,
        "parameters": [
          {
            "name": "id" ,
            "in": "path" ,
            "required": true ,
            "schema": {
              "type": "integer" ,
              "format": "int64"
            }
          }
        ] ,
        "responses": {
          "400": {
            "description": "Bad Request" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "401": {
            "description": "Unauthorized" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "200": {
            "description": "OK" ,
            "content": {
              "*/*": {
                "schema": {
                  "type": "array" ,
                  "items": {
                    "$ref": "#/components/schemas/PersonneImpliquee"
                  }
                }
              }
            }
          }
        }
      }
    } ,
...
    "/api/v1/assurance": {
      "get": {
        "tags": [
          "assurance-controller"
        ] ,
        "operationId": "assuranceList" ,
        "responses": {
          "400": {
            "description": "Bad Request" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "401": {
            "description": "Unauthorized" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "200": {
            "description": "OK" ,
            "content": {
              "*/*": {
                "schema": {
                  "type": "array" ,
                  "items": {
                    "$ref": "#/components/schemas/Assurance"
                  }
                }
              }
            }
          }
        }
      }
    } ,
    "/api/v1/assurance/{id}": {
      "get": {
        "tags": [
          "assurance-controller"
        ] ,
        "operationId": "details" ,
        "parameters": [
          {
            "name": "id" ,
            "in": "path" ,
            "required": true ,
            "schema": {
              "type": "integer" ,
              "format": "int64"
            }
          }
        ] ,
        "responses": {
          "400": {
            "description": "Bad Request" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "401": {
            "description": "Unauthorized" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "200": {
            "description": "OK" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Assurance"
                }
              }
            }
          }
        }
      }
    } ,
    "/api/v1/accident/{id}": {
      "get": {
        "tags": [
          "accident-controller"
        ] ,
        "operationId": "details_1" ,
        "parameters": [
          {
            "name": "id" ,
            "in": "path" ,
            "required": true ,
            "schema": {
              "type": "integer" ,
              "format": "int64"
            }
          }
        ] ,
        "responses": {
          "400": {
            "description": "Bad Request" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "401": {
            "description": "Unauthorized" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "200": {
            "description": "OK" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Accident"
                }
              }
            }
          }
        }
      }
    } ,
...
    "/api/v1/personne-impliquee/delete/{id}": {
      "delete": {
        "tags": [
          "personne-impliquee-controller"
        ] ,
        "operationId": "delete_3" ,
        "parameters": [
          {
            "name": "id" ,
            "in": "path" ,
            "required": true ,
            "schema": {
              "type": "integer" ,
              "format": "int64"
            }
          }
        ] ,
        "responses": {
          "400": {
            "description": "Bad Request" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "401": {
            "description": "Unauthorized" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "200": {
            "description": "OK"
          }
        }
      }
    } ,
    "/api/v1/assurance/delete/{id}": {
      "delete": {
        "tags": [
          "assurance-controller"
        ] ,
        "operationId": "delete_4" ,
        "parameters": [
          {
            "name": "id" ,
            "in": "path" ,
            "required": true ,
            "schema": {
              "type": "integer" ,
              "format": "int64"
            }
          }
        ] ,
        "responses": {
          "400": {
            "description": "Bad Request" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "401": {
            "description": "Unauthorized" ,
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponse"
                }
              }
            }
          } ,
          "200": {
            "description": "OK"
          }
        }
      }
    } ,
...
  } ,
  "components": {
    "schemas": {
      "HttpResponse": {
        "type": "object" ,
        "properties": {
          "timestamp": {
            "type": "string" ,
            "format": "date-time"
          } ,
          "statusCode": {
            "type": "integer" ,
            "format": "int32"
          } ,
          "httpStatus": {
            "type": "string" ,
            "enum": [
              "100 CONTINUE" ,
              "101 SWITCHING_PROTOCOLS" ,
              "102 PROCESSING" ,
              "103 CHECKPOINT" ,
              "200 OK" ,
              "201 CREATED" ,
              "202 ACCEPTED" ,
              "203 NON_AUTHORITATIVE_INFORMATION" ,
              "204 NO_CONTENT" ,
              "205 RESET_CONTENT" ,
              "206 PARTIAL_CONTENT" ,
              "207 MULTI_STATUS" ,
              "208 ALREADY_REPORTED" ,
              "226 IM_USED" ,
              "300 MULTIPLE_CHOICES" ,
              "301 MOVED_PERMANENTLY" ,
              "302 FOUND" ,
              "302 MOVED_TEMPORARILY" ,
              "303 SEE_OTHER" ,
              "304 NOT_MODIFIED" ,
              "305 USE_PROXY" ,
              "307 TEMPORARY_REDIRECT" ,
              "308 PERMANENT_REDIRECT" ,
              "400 BAD_REQUEST" ,
              "401 UNAUTHORIZED" ,
              "402 PAYMENT_REQUIRED" ,
              "403 FORBIDDEN" ,
              "404 NOT_FOUND" ,
              "405 METHOD_NOT_ALLOWED" ,
              "406 NOT_ACCEPTABLE" ,
              "407 PROXY_AUTHENTICATION_REQUIRED" ,
              "408 REQUEST_TIMEOUT" ,
              "409 CONFLICT" ,
              "410 GONE" ,
              "411 LENGTH_REQUIRED" ,
              "412 PRECONDITION_FAILED" ,
              "413 PAYLOAD_TOO_LARGE" ,
              "413 REQUEST_ENTITY_TOO_LARGE" ,
              "414 URI_TOO_LONG" ,
              "414 REQUEST_URI_TOO_LONG" ,
              "415 UNSUPPORTED_MEDIA_TYPE" ,
              "416 REQUESTED_RANGE_NOT_SATISFIABLE" ,
              "417 EXPECTATION_FAILED" ,
              "418 I_AM_A_TEAPOT" ,
              "419 INSUFFICIENT_SPACE_ON_RESOURCE" ,
              "420 METHOD_FAILURE" ,
              "421 DESTINATION_LOCKED" ,
              "422 UNPROCESSABLE_ENTITY" ,
              "423 LOCKED" ,
              "424 FAILED_DEPENDENCY" ,
              "425 TOO_EARLY" ,
              "426 UPGRADE_REQUIRED" ,
              "428 PRECONDITION_REQUIRED" ,
              "429 TOO_MANY_REQUESTS" ,
              "431 REQUEST_HEADER_FIELDS_TOO_LARGE" ,
              "451 UNAVAILABLE_FOR_LEGAL_REASONS" ,
              "500 INTERNAL_SERVER_ERROR" ,
              "501 NOT_IMPLEMENTED" ,
              "502 BAD_GATEWAY" ,
              "503 SERVICE_UNAVAILABLE" ,
              "504 GATEWAY_TIMEOUT" ,
              "505 HTTP_VERSION_NOT_SUPPORTED" ,
              "506 VARIANT_ALSO_NEGOTIATES" ,
              "507 INSUFFICIENT_STORAGE" ,
              "508 LOOP_DETECTED" ,
              "509 BANDWIDTH_LIMIT_EXCEEDED" ,
              "510 NOT_EXTENDED" ,
              "511 NETWORK_AUTHENTICATION_REQUIRED"
            ]
          } ,
          "reason": {
            "type": "string"
          } ,
          "message": {
            "type": "string"
          }
        }
      } ,
      "Prefecture": {
        "type": "object" ,
        "properties": {
          "id": {
            "type": "integer" ,
            "format": "int64"
          } ,
          "nom": {
            "type": "string"
          } ,
          "nombreHabitant": {
            "type": "integer" ,
            "format": "int64"
          } ,
          "region": {
            "$ref": "#/components/schemas/Region"
          } ,
          "createdAt": {
            "type": "string" ,
            "format": "date-time"
          } ,
          "updateAt": {
            "type": "string" ,
            "format": "date-time"
          }
        }
      } ,
      "Region": {
        "type": "object" ,
        "properties": {
          "id": {
            "type": "integer" ,
            "format": "int64"
          } ,
          "nom": {
            "type": "string"
          } ,
          "nombreHabitant": {
            "type": "integer" ,
            "format": "int64"
          } ,
          "createdAt": {
            "type": "string" ,
            "format": "date-time"
          } ,
          "updateAt": {
            "type": "string" ,
            "format": "date-time"
          }
        }
      } ,
      "Ville": {
        "type": "object" ,
        "properties": {
          "id": {
            "type": "integer" ,
            "format": "int64"
          } ,
          "nom": {
            "type": "string"
          } ,
          "nombreHabitant": {
            "type": "integer" ,
            "format": "int64"
          } ,
          "prefecture": {
            "$ref": "#/components/schemas/Prefecture"
          } ,
          "createdAt": {
            "type": "string" ,
            "format": "date-time"
          } ,
          "updateAt": {
            "type": "string" ,
            "format": "date-time"
          }
        }
      } ,
      "Accident": {
        "type": "object" ,
        "properties": {
          "id": {
            "type": "integer" ,
            "format": "int64"
          } ,
          "dateAccident": {
            "type": "string" ,
            "format": "date"
          } ,
          "heure": {
            "type": "string" ,
            "format": "date-time"
          } ,
          "zone": {
            "type": "string"
          } ,
          "quartier": {
            "type": "string"
          } ,
          "intersectionAccident": {
            "type": "string"
          } ,
          "pointRepere": {
            "type": "string"
          } ,
          "latitude": {
            "type": "number" ,
            "format": "float"
          } ,
          "longitude": {
            "type": "number" ,
            "format": "float"
          } ,
          "route": {
            "type": "string"
          } ,
          "troncon": {
            "type": "string"
          } ,
          "circonstanceResume": {
            "type": "string"
          } ,
          "typeJour": {
            "type": "string"
          } ,
          "codificationCauses": {
            "type": "string"
          } ,
          "lumiere": {
            "type": "string"
          } ,
          "profilLieu": {
            "type": "string"
          } ,
          "conditionAtmospherique": {
            "type": "string"
          } ,
          "isMarquageSol": {
            "type": "boolean"
          } ,
          "traceSol": {
            "type": "string"
          } ,
          "etatChaussee": {
            "type": "string"
          } ,
          "controleCarrefour": {
            "type": "string"
          } ,
          "surfaceAtteinte": {
            "minimum": 0 ,
            "type": "number" ,
            "format": "double"
          } ,
          "nombreMorts": {
            "minimum": 0 ,
            "type": "integer" ,
            "format": "int32"
          } ,
          "nombreVehiculesImpliques": {
            "minimum": 0 ,
            "type": "integer" ,
            "format": "int32"
          } ,
          "nombrePanneauxAtteints": {
            "minimum": 0 ,
            "type": "integer" ,
            "format": "int32"
          } ,
          "nombreBalisesAtteintes": {
            "minimum": 0 ,
            "type": "integer" ,
            "format": "int32"
          } ,
          "nombreGlissieresAtteintes": {
            "minimum": 0 ,
            "type": "integer" ,
            "format": "int32"
          } ,
          "nombrePoteauxAtteints": {
            "minimum": 0 ,
            "type": "integer" ,
            "format": "int32"
          } ,
          "nombreGardeFouAtteints": {
            "minimum": 0 ,
            "type": "integer" ,
            "format": "int32"
          } ,
          "nombreOuvrageBetonAtteints": {
            "minimum": 0 ,
            "type": "integer" ,
            "format": "int32"
          } ,
          "autreDegats": {
            "type": "string"
          } ,
          "images": {
            "type": "array" ,
            "items": {
              "$ref": "#/components/schemas/Image"
            }
          } ,
          "ville": {
            "$ref": "#/components/schemas/Ville"
          } ,
          "typeRoute": {
            "$ref": "#/components/schemas/TypeRoute"
          } ,
          "classificationRoute": {
            "$ref": "#/components/schemas/ClassificationRoute"
          } ,
          "user": {
            "$ref": "#/components/schemas/User"
          } ,
          "createdAt": {
            "type": "string" ,
            "format": "date-time"
          } ,
          "updateAt": {
            "type": "string" ,
            "format": "date-time"
          }
        }
      } ,
      "Assurance": {
        "type": "object" ,
        "properties": {
          "id": {
            "type": "integer" ,
            "format": "int64"
          } ,
          "nom": {
            "type": "string"
          }
        }
      } ,
      "ClassificationRoute": {
        "type": "object" ,
        "properties": {
          "id": {
            "type": "integer" ,
            "format": "int64"
          } ,
          "nom": {
            "type": "string"
          }
        }
      } ,
      "Image": {
        "type": "object" ,
        "properties": {
          "id": {
            "type": "integer" ,
            "format": "int64"
          } ,
          "nom": {
            "type": "string"
          } ,
          "url": {
            "type": "string"
          }
        }
      } ,
      "Role": {
        "type": "object" ,
        "properties": {
          "id": {
            "type": "integer" ,
            "format": "int64"
          } ,
          "nom": {
            "type": "string"
          }
        }
      } ,
      "TypeRoute": {
        "type": "object" ,
        "properties": {
          "id": {
            "type": "integer" ,
            "format": "int64"
          } ,
          "nom": {
            "type": "string"
          }
        }
      } ,
      "User": {
        "type": "object" ,
        "properties": {
          "id": {
            "type": "integer" ,
            "format": "int64"
          } ,
          "userMatricule": {
            "type": "string"
          } ,
          "nom": {
            "type": "string"
          } ,
          "password": {
            "type": "string" ,
            "writeOnly": true
          } ,
          "lastLoginDate": {
            "type": "string" ,
            "format": "date-time"
          } ,
          "lastLoginDateDisplay": {
            "type": "string" ,
            "format": "date-time"
          } ,
          "joinDate": {
            "type": "string" ,
            "format": "date-time"
          } ,
          "isActive": {
            "type": "boolean"
          } ,
          "isNotLocked": {
            "type": "boolean"
          } ,
          "token": {
            "type": "string"
          } ,
          "roles": {
            "type": "array" ,
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          } ,
          "tokenValid": {
            "type": "boolean"
          }
        }
      } ,
      "Vehicule": {
        "type": "object" ,
        "properties": {
          "id": {
            "type": "integer" ,
            "format": "int64"
          } ,
          "immatriculation": {
            "type": "string"
          } ,
          "paysImmatriculation": {
            "type": "string"
          } ,
          "genre": {
            "type": "string"
          } ,
          "isPremierChoc": {
            "type": "boolean"
          } ,
          "debutCirculation": {
            "type": "string"
          } ,
          "isControleTechValid": {
            "type": "boolean"
          } ,
          "isAssuranceValid": {
            "type": "boolean"
          } ,
          "etatGeneral": {
            "type": "string"
          } ,
          "etatPneus": {
            "type": "string"
          } ,
          "etatFeux": {
            "type": "string"
          } ,
          "typeChargement": {
            "type": "string"
          } ,
          "volumeChargement": {
            "type": "string"
          } ,
          "createdAt": {
            "type": "string" ,
            "format": "date-time"
          } ,
          "updateAt": {
            "type": "string" ,
            "format": "date-time"
          } ,
          "accident": {
            "$ref": "#/components/schemas/Accident"
          } ,
          "assurance": {
            "$ref": "#/components/schemas/Assurance"
          }
        }
      } ,
...
      "TypePersonne": {
        "type": "object" ,
        "properties": {
          "id": {
            "type": "integer" ,
            "format": "int64"
          } ,
          "nom": {
            "type": "string"
          } ,
          "createdAt": {
            "type": "string" ,
            "format": "date-time"
          } ,
          "updateAt": {
            "type": "string" ,
            "format": "date-time"
          }
        }
      } ,
      "Profession": {
        "type": "object" ,
        "properties": {
          "id": {
            "type": "integer" ,
            "format": "int64"
          } ,
          "nom": {
            "type": "string"
          } ,
          "description": {
            "type": "string"
          } ,
          "createdAt": {
            "type": "string" ,
            "format": "date-time"
          } ,
          "updateAt": {
            "type": "string" ,
            "format": "date-time"
          }
        }
      } ,
      "PersonneImpliquee": {
        "type": "object" ,
        "properties": {
          "id": {
            "type": "integer" ,
            "format": "int64"
          } ,
          "cni": {
            "type": "string"
          } ,
          "nom": {
            "type": "string"
          } ,
          "prenom": {
            "type": "string"
          } ,
          "age": {
            "minimum": 0 ,
            "type": "integer" ,
            "format": "int32"
          } ,
          "sexe": {
            "type": "string"
          } ,
          "nationalite": {
            "type": "string"
          } ,
          "lesion": {
            "type": "string"
          } ,
          "protection": {
            "type": "string"
          } ,
          "gravite": {
            "type": "string"
          } ,
          "alcoolemie": {
            "type": "boolean"
          } ,
          "stupefiant": {
            "type": "boolean"
          } ,
          "typePermis": {
            "type": "string"
          } ,
          "numeroPermis": {
            "type": "string"
          } ,
          "anneePermis": {
            "type": "string"
          } ,
          "paysPermis": {
            "type": "string"
          } ,
          "createdAt": {
            "type": "string" ,
            "format": "date-time"
          } ,
          "updateAt": {
            "type": "string" ,
            "format": "date-time"
          } ,
          "profession": {
            "$ref": "#/components/schemas/Profession"
          } ,
          "typePersonne": {
            "$ref": "#/components/schemas/TypePersonne"
          } ,
          "vehicule": {
            "$ref": "#/components/schemas/Vehicule"
          }
        }
      } ,
...
    }
  }
}